+++ arika/src/earth/mod.rs
@@ -23,9 +23,11 @@
 
 pub mod ellipsoid;
 pub mod eop;
+pub mod fk5;
 pub mod geodetic;
 pub mod iau2006;
 pub mod rotation;
+pub mod teme;
 pub mod topocentric;
 pub mod transform;